Add singly and doubly-linked list implementations
This commit is the initial commit and, as such, initialises the
repository. It also adds working implementations of linked-lists,
both singly and doubly-linked varieties. Basic documentation exists
in the form of READMEs and Markdown files in the `docs/` directories
of each respective implementation directory (e.g.
`lists/dlist/docs/dlist.md`).