Make data fields for linked lists type-agnostic
This commit makes the `data` field in both singly-linked lists and
doubly-linked lists type-agnostic via the use of `void` pointers.

The previous definition of the `Node` structure was along the lines of:

`struct Node
{
  int data;
  struct Node* next;
  struct Node* prev; // only for doubly-linked lists
}`

but now resembles this:

`struct Node
{
  void* data;
  struct Node* next;
  struct Node* prev; // only for doubly-linked lists
}`

This improves abstraction and portability of code. Note that all
traversal functions still consider the `data` field to be of type `int`
and cast the `data` field accordingly.