Add stack and change doubly-linked list behaviour
This commit adds a stack implementation, located in the `stack`
directory. The stack is implemented via a doubly-linked list (i.e. the
code in `dlist`) and thus, depends on the header file `dlist.h`.

Due to the stack implementation, it became apparent that poor choices
were made while implementing the doubly-linked list. In `dlist.h`, the
function `insertAfter` inserts a node in the linked list, *after* the
specified node. It accepts a pointer to the existing node (i.e. the
node to insert the new node after) and a pointer to the new node (i.e.
the node to be added to the list). If the existing node is non-existant
(i.e. `target_node == NULL`, where `target_node` is of type
`struct Node*`), `insertAfter` returned `EXIT_FAILURE` (the C library
macro defined in `<stdlib.h>`). This behaviour has now been changed to
simply insert the new node, but set its previous node to `NULL`. This
change occurs on line 63 of `dlist.h`.